refactor(message): split boundaries and normalize comments
