+++ CMakeLists.txt
@@ -73,6 +73,11 @@ file(GLOB ${PROJECT_NAME}_MIDDLEWARE_SUBDIR_SOURCES CONFIGURE_DEPENDS
 )
 list(SORT ${PROJECT_NAME}_MIDDLEWARE_SUBDIR_SOURCES)
 
+file(GLOB ${PROJECT_NAME}_MIDDLEWARE_NESTED_SUBDIR_SOURCES CONFIGURE_DEPENDS
+     "${CMAKE_CURRENT_SOURCE_DIR}/src/middleware/*/*/*.cpp"
+)
+list(SORT ${PROJECT_NAME}_MIDDLEWARE_NESTED_SUBDIR_SOURCES)
+
 target_sources(
   ${PROJECT_NAME}
   PRIVATE ${${PROJECT_NAME}_SOURCES}
@@ -84,6 +89,7 @@ target_sources(
   PRIVATE ${${PROJECT_NAME}_STRUCTURE_SOURCES}
   PRIVATE ${${PROJECT_NAME}_MIDDLEWARE_SOURCES}
   PRIVATE ${${PROJECT_NAME}_MIDDLEWARE_SUBDIR_SOURCES}
+  PRIVATE ${${PROJECT_NAME}_MIDDLEWARE_NESTED_SUBDIR_SOURCES}
 )
 
 include(${CMAKE_CURRENT_SOURCE_DIR}/src/core/CMakeLists.txt)
