+++ src/middleware/message/packet/packet.cpp
@@ -1,10 +1,12 @@
+#include "packet.hpp"
+
 #include "libxr_mem.hpp"
-#include "message.hpp"
 
 using namespace LibXR;
 
 void Topic::PackedDataHeader::SetDataLen(uint32_t len)
 {
+  // 大端写入 24 位长度
   data_len_raw[0] = static_cast<uint8_t>(len >> 16);
   data_len_raw[1] = static_cast<uint8_t>(len >> 8);
   data_len_raw[2] = static_cast<uint8_t>(len);
@@ -19,6 +21,7 @@ uint32_t Topic::PackedDataHeader::GetDataLen() const
 
 void Topic::PackedDataHeader::SetTimestamp(MicrosecondTimestamp timestamp)
 {
+  // 大端写入时间戳
   uint64_t value = static_cast<uint64_t>(timestamp);
   for (size_t i = 0; i < sizeof(timestamp_us_raw); i++)
   {
@@ -50,6 +53,7 @@ void Topic::PackData(uint32_t topic_name_crc32, RawData buffer,
   pack->raw.header_.SetTimestamp(timestamp);
   pack->raw.header_.pack_header_crc8 =
       CRC8::Calculate(&pack->raw, sizeof(PackedDataHeader) - sizeof(uint8_t));
+  // 尾 CRC8 紧跟 payload
   uint8_t* crc8_pack = reinterpret_cast<uint8_t*>(
       reinterpret_cast<uint8_t*>(pack) + PACK_BASE_SIZE + data.size_ - sizeof(uint8_t));
   *crc8_pack = CRC8::Calculate(pack, PACK_BASE_SIZE - sizeof(uint8_t) + data.size_);
